Fix negative y-axis in TrackerView
Any unpointed stories in the current iteration contribute -1 points to
their state count. This comes directly from the Pivotal Tracker API.

If there are no (or not enough) pointed stories in a state that has
unpointed stories, then that state will have a negative point value.
This means that the y-axis will have a negative minimum value and that
the other states will have a bar from the negative minimum value up to
zero (or greater as the case may be).

This commit sanitizes the input from Pivotal Tracker and only considers
estimated stories for each state count.

[Fixes #77045938]