CircleCI projects are not being updated via webhooks. [#94687742]
This bug was caused by two problems:

1. The CircleCiProject did not have a webhook_payload method
2. The CircleCiPayload was incorrectly parsing the webhook payload that
webhooks was sending.

I also did a refactoring so that every class that extends Project
automatically gets a webhook_payload method that calls the class'
fetch_payload method.  I did this because as far as I could see, every
Project should have the former and every class already has the latter
and the implementation for webhook_payload is almost always the same
as fetch_payload.