Prune 2014 interactions from VCR cassettes
The current VCR setup is trying to re-record cassettes every
18 months with the {vcr: {re_record_interval: 18.months}} option,
but that date is compared against the *oldest* interaction
in the cassette, which is still from 2014.

It seems like when the casettes are 're-recorded' it doesn't
clear out the old content, but that makes the :re_record_interval
option useless because once you've hit the re_record interval
it will keep re-recording on every run forever. Not sure how
to fix this other than making it a more manual process.

Additionally, the test fails today if VCR is allowed to
re-record, probably something about Sundays or time zones.