Fix regression causing the 'Refresh' button on the editor to create new projects
This was caused by the Rails 4.1 upgrade.

The code was creating a new (non-persisted) Project instance, which had
the same attributes as a project in the database except it lacked an id.

In Rails 4.0, when creating a payload_log_entry associated with this
project (`log = project.payload_log_entries.build; log.save!`) the
parent project record would remain unpersisted.

Rails 4.1, for whatever reason, attempts to save the parent project record as
if it's a new project. Hammering on the 'Refresh' button to make more requests
would thereby make a new project every time.

The fix here is twofold:
* When finding a project record in the controller, always find it from the
database if the project has been persisted, instead of trying to create a new
in-memory record.
* Only try to save the associated PayloadLogEntry if the associated project was persisted.